+++ core/store/orm/schema.go
@@ -73,6 +73,7 @@ type ConfigSchema struct {
 	GasUpdaterBlockHistorySize                uint16          `env:"GAS_UPDATER_BLOCK_HISTORY_SIZE"`
 	GasUpdaterEnabled                         bool            `env:"GAS_UPDATER_ENABLED"`
 	GasUpdaterTransactionPercentile           uint16          `env:"GAS_UPDATER_TRANSACTION_PERCENTILE" default:"60"`
+	GinMode                                   string          `env:"GIN_MODE"`
 	GlobalLockRetryInterval                   models.Duration `env:"GLOBAL_LOCK_RETRY_INTERVAL" default:"1s"`
 	HTTPServerWriteTimeout                    time.Duration   `env:"HTTP_SERVER_WRITE_TIMEOUT" default:"10s"`
 	HeadTimeBudget                            time.Duration   `env:"HEAD_TIME_BUDGET"`